test if notion image expire triggers rebuild
