+++ components/NotionTextBlock.tsx
@@ -1,0 +1,31 @@
+export function Text({ text }: { text: any }) {
+  if (!text) {
+    return null
+  }
+  return text.map((value: any) => {
+    const {
+      annotations: { bold, code, color, italic, strikethrough, underline },
+      text,
+    } = value
+    return (
+      <span
+        className={[
+          bold ? 'font-bold' : '',
+          code ? 'font-mono text-sm px-1 bg-light-300 dark:bg-dark-800 rounded' : '',
+          italic ? 'italic' : '',
+          strikethrough ? 'strike' : '',
+          underline ? 'underline' : '',
+        ].join(' ')}
+        style={color !== 'default' ? { color } : {}}
+      >
+        {text.link ? (
+          <a href={text.link.url} target="_blank" rel="noopener noreferrer">
+            {text.content}
+          </a>
+        ) : (
+          text.content
+        )}
+      </span>
+    )
+  })
+}
