+++ next.config.js
@@ -4,7 +4,7 @@ const WindiCSSWebpackPlugin = require('windicss-webpack-plugin')
 module.exports = {
   reactStrictMode: true,
   images: {
-    domains: ['s3-us-west-2.amazonaws.com']
+    domains: ['s3-us-west-2.amazonaws.com', 's3.us-west-2.amazonaws.com']
   },
   webpack(config) {
     config.plugins.push(new WindiCSSWebpackPlugin())
